[Autofill] Move probable-form-submission detection to the browser.
This CL moves the probable-form-submission detection from the renderer
process to the browser process. The new behaviour is active iff
kAutofillEnableProbableFormSubmissionInBrowser is enabled.

The current state is that probable form submission is detected in
FormTracker::DidStartNavigation(), which forwards the event to
AutofillAgent::OnProbablyFormSubmitted(), which then notifies
ContentAutofillDriver::FormSubmitted(). The 'submitted form' is the
one returned by AutofillAgent::GetSubmittedForm().

This CL moves the event detection to the newly added
ContentAutofillDriverFactory::DidStartNavigation(), which calls the
newly added ContentAutofillDriver::ProbablyFormSubmitted(). The
'submitted form' should again be the one returned by
AutofillAgont::GetSubmittedForm(). For the browser process to have
access to that form, AutofillAgent now proactively sends it to the
browser process by calling SendPotentiallySubmittedFormToBrowser().

See the referenced bugs for further details.

Bug: 1113740, 1117451
Change-Id: I46be9e8b7058f4a309e3e9c64f34ca4d8852e641
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2332622
Commit-Queue: Christoph Schwering <schwering@google.com>
Reviewed-by: Matthias Körber <koerber@google.com>
Reviewed-by: Mike West <mkwst@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800092}