[PwdCheckAndroid] Add dialog to view a compromised credential
This CL adds a dialog that shows a compromised credential. The dialog
is shown when the ¨View password¨ entry is clicked on the More menu of
a compromised credential. It also adds a test to verify the close
button of the dialog works correctly.

Screenshot in the linked bug.

Bug: 1117502, 1092444
Change-Id: Iff5784b3e78a3f58072c7e7944f85f6d93b6820a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363750
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Eleonora Rocchi <erocchi@google.com>
Cr-Commit-Position: refs/heads/master@{#800107}