[WebLayer Shell] Make URL TextView clickable again.
I had removed the clicklistener on the TextView for the demo app
here:
https://chromium-review.googlesource.com/c/chromium/src/+/2362817

since I was setting it on the parent view. Turns out, to make the child
TextView clicks go up to the parent, I'll need to mark it non-clickable in the
layout xml, which we don't want to do for production.

Adding the click listener explicitly on the TextView fixes this.

Change-Id: I7feaf35a436abb4ca4313f94b12e448c74f58562
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2366900
Reviewed-by: Bo <boliu@chromium.org>
Commit-Queue: Mugdha Lakhani <nator@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800109}