Remove kCommittedMultipleRealLoads from FrameLoaderStateMachine
Its purpose is to avoid creating spurious history entries in the case
where we replace the initial empty document with a different
about:blank document (which the spec says not to do, but I think every
major browser does). Explicitly track this case with a bit on
FrameLoader for whether we've ever loaded a non-empty document.

Change-Id: I2bd6d759ccf6f43d14feddeac1f72d2f871a80cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354995
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Commit-Queue: Nate Chapin <japhet@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800149}