[PwdCheckAndroid] Add credentials sorting in java
When a credential from the list changes (is added in a new check or
removed) the entire list has to change. However, in order to keep the UI
stable, the list should maintain its order, which has to be as follows:

1. Phished credentials appear before leaked credentials.
2. Compromised credentials found in previous checks should appear
in descending order of creation time (most recent first).
3. Compromised credentials found in the current check (and all checks
ran while the list UI is open) have to be appended to the list.
This means that they need to be ordered in ascending order of creation
time.

Bug: 1102025, 1092444
Change-Id: I9e262791227454065fd556c0974b12b43fc9843a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362609
Commit-Queue: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800156}