Show a different text for long-term support builds
When the DeviceReleaseLtsTag policy is set, display a different text in
Chrome OS settings to tell the user that they are currently using a
long-term support (LTS) build.

Additonally add the LTS-specific text to "Channel" entry when Build
Details are copied to clipboard.

Note: For the LTS pilot, we only check whether the DeviceReleaseLtsTag
policy is set and do not consider the value it is set to. This will
change for future LTS releases once the exact value is decided on.

Bug: 1105495
Change-Id: Ic1c11e14e690de8ab900337c81487d931ba9fd32
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2355925
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Victor-Gabriel Savu <vsavu@google.com>
Commit-Queue: Jana Grill <janagrill@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800217}