Include lock URL and origin-keyed bit in SiteInfo::GetDebugString()
This will make test failures involving SiteInfo comparisons more
readable, and it will also include the process lock and origin-keyed
information in site URL crash keys logged on renderer kills.
Similarly to https://crrev.com/c/2364816, the process lock URL is
appended only when it's different from site URL, since the two are the
same in the common case.

A common use of different site and lock URLs is hosted apps, where the
site URL is of form chrome-extension://<32-char-id>/.  With that in
mind, the SiteInfo crash keys are increased in length from 64 to 256,
since 64 isn't enough to cover the 52-char chrome-extension site URL
together with a lock URL.

Bug: 1105994, 1067389
Change-Id: Ic747a9ca9244e4da2eebbeacbd499f7a8a439748
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2364579
Commit-Queue: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Aaron Colwell <acolwell@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800341}