Add histograms_index file to record all available histograms xmls
This histograms_index file provides a convenient way for a job to know
all of the histogram files without having to search for them. As we
split up histograms, we need a way to locate all available histograms
xmls and thus we need this index file.

This cl also adds a PRESUBMIT check to make sure this index file is
up-to-date.

Bug: 1116807
Change-Id: Ifc55e421092448b5cecaa573f02e651a098428bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2359234
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Brian White <bcwhite@chromium.org>
Commit-Queue: Weilun Shi <sweilun@chromium.org>
Auto-Submit: Weilun Shi <sweilun@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800357}