Add field trial config to disable specific heavy ad thresholds
What: Add a feaature param to selectively disable the network or CPU
limits.

Why: We may want to run the intervention in different modes to
determine impact.

How: Add a check after an ad has determined to be heavy that filters
based on the feature param.

This change also cleans up some logic by returning an enum instead
of bool in MaybeTriggerHeavyAdIntervention(), and using the new
HeavyAdAction value to control triggering of interventions.

This also removes a mode where the intervention could be enabled
without reporting, which never needs to be used.

Change-Id: Ie7aefdd77edfee5b2d4463aab3e1d1790a0931b2
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363190
Commit-Queue: John Delaney <johnidel@chromium.org>
Reviewed-by: Eric Robinson <ericrobinson@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800396}