[Launcher impressions] Use metadata_type member for recording types.
Right now we record type metrics like this:
 - user action received by ash
 - ash sends result id to chrome
 - chrome re-associates id with a ChromeSearchResult object
 - we log the type associated with the ChromeSearchResult object

However the re-association step is subject to a race condition related
to whether the search result objects have been cleared from the cache.

This CL changes the process to be:
 - user action received by ash
 - ash sends an AppListNotifier::Result object to chrome, containing
   id and type
 - chrome logs type directly.

Note: the result id string in AppListNotifier::Result is now unused on
the chrome side, but we're planning to use it for other
impressions-related work soon.

Bug: 1076270
Change-Id: I699f293139ac7b89cbcfe92aa9c5af017b033a48
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2365572
Commit-Queue: Tony Yeoman <tby@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Rachel Wong <wrong@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800441}