[TTS] Add an IPH to enable for improving translations
Adds a new In Product Help bubble that suggests that the user
should enable access to page content in order to improve translations.

This includes handling of the IPH, but not any triggering of the IPH
or noting that the expected user response event has occurred.

Following instructions here:
https://chromium.googlesource.com/chromium/src/+/refs/heads/master/docs/translation_screenshots.md

BUG=1120080

Change-Id: Ib929b40d9b0954bfba5ed80ba02d3913ec30f1f0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2368472
Commit-Queue: Theresa  <twellington@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Auto-Submit: Donn Denman <donnd@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800459}