[signin] Supports  OAuth outage in Dice.
This is the client side implementation of go/gaiafe-dice-fallback.
The Gaia server will send parameter no_authorization_code in the Dice
header, when there is a OAuth outage.

When this happens and the feature kSupportOAuthOutageInDice  is enabled,
Chrome disable the reconcilor temporarily, so that the user can stay
signed in on the web.
After some time, the reconcilor is restarted, and at this point the user
is signed out, and will need to sign in again.

Also added unit tests.

Bug: 1115247
Change-Id: Ib71f55f26f50c185175a1abeec58edd4d7d72605
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2362757
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Diana Zagidullina <dizg@google.com>
Cr-Commit-Position: refs/heads/master@{#800512}