[Android][Mfill] Implement AllPasswordsBottomSheetView
This Cl adds |AllPasswordsBottomSheetView| which implements
|BottomSheetContent| and chnge its visibility state from the controller.
Adds a junit test.

This Cl skips adding screenshots for string resources. The UI is in
early stages and taking a screenshot for it does not help for
translation.
Will be added when the UI gives helpful context.

Bug: 1104132
Skip-Translation-Screenshots-Check: True
Change-Id: If6be6f9dbea93489b69b367d4d997cec595fb7d1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2320171
Reviewed-by: Ioana Pandele <ioanap@chromium.org>
Reviewed-by: Friedrich [CET] <fhorschig@chromium.org>
Commit-Queue: Reda Tawfik <redatawfik@noogler.google.com>
Cr-Commit-Position: refs/heads/master@{#800517}