Align "Manage extensions" text and icon
The extension icon images are 28x28 to as they contain space for badging
the icon with "click to script". They are still visually 16x16 icons.
This change adds the same amount of space (but not for badging) around
the actually-16x16 settings icon to make icon and text aligned with
other items in this menu.

Bug: 1050601
Change-Id: I51c9aa81ab05fd53cf3b3702e73c543f257a1960
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2368927
Commit-Queue: Peter Boström <pbos@chromium.org>
Reviewed-by: Elly Fong-Jones <ellyjones@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800560}