Recover icons on multiple monitors when release the dragged icon.
With multiple monitors, when launch an app unpin on the shelf, and drag
the icon for the launched app, the icon is ripped off from the shelf on
all monitors, because the icon opacity is set as 0.0f (transparent):
https://source.chromium.org/chromium/chromium/src/+/master:ash/shelf/shelf_view.cc;l=2155

However, when release the button to restore the icon back to the Shelf,
only the icon on 1 monitor pulled by the mouse is recovered because the
opacity for the icon is restored to 1.0, not transparent by the drag
view:
drag_view_->layer()->SetOpacity(1.0f);

Well other icons on other monitors are not recovered. So modify the drag
rip off handling to recover other icons on other monitors when release
the mouse for the drag icon.

BUG=1119698

Change-Id: I49439a7c02c18b0875c53772bd0081a016ab720d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367261
Commit-Queue: Nancy Wang <nancylingwang@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800579}