[mini_installer] Move the fallback reporting to InstallerError.
https://crrev.com/798844 attempted to send up a signal regarding whether
or not the fallback temp directory was used. Unfortunately, Omaha does
not include the ExtraCode1 value for successful installs. It does,
however, send up success HRESULTS in the InstallerError value. This CL,
therefore, switches to that. The two values to look out for are
0x00044001 and 0x00044002 (two success HRESULTS in FACILITY_ITF that
otherwise appear to be unused).

BUG=516207,1100280

Change-Id: I319c5dbe7162543a71465238c5320719225fa1f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2368353
Auto-Submit: Greg Thompson <grt@chromium.org>
Commit-Queue: S. Ganesh <ganesh@chromium.org>
Reviewed-by: S. Ganesh <ganesh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800584}