[CCTToSFRE] Change the FRE layout
Change the FRE UI to better adapt to the transition from loading to ToS.

Layout preview: go/clank-fre-ui-preview

This CL also changes the visibility of UMA / ToS component to View.GONE
when they are hidden. This will free up the space so that the screen
will not be scrollable on blank content.

Bug: 1108558, 1117352
Change-Id: I3f8616bd5810da71afa8b6e40dea2db92e89013e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2352033
Commit-Queue: Wenyu Fu <wenyufu@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800591}