arc: Wait longer for PDF to render when printing
PrintSessionImpl checks every 100ms for a subframe that has a PDF
extension loaded. When it finds the subframe and sees that it's done
loading, it waits 100ms for the PDF to finish rendering and then
launches print preview. Recent changes increased the amount of time it
takes for the PDF to render, so slower devices are launching print
preview too early.

Increase the time to wait for the PDF to render from 100ms to 500ms to
ensure slower devices can render the PDF before launching print preview.

Bug: b:165927343
Test: arc.Print
Change-Id: If3e20fc2ffa3feb74e112e1e88c8248f164e80bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2369613
Reviewed-by: Lei Zhang <thestig@chromium.org>
Commit-Queue: Jesse Schettler <jschettler@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800700}