xkblayoutengine: fix memory leak by unrefing keymap from buffer.
Wayland may supply xkb layout engine with a keymap that it creates
from a buffer. However, when I ran events_unittests with asan,
it noticed a memory leak, which was caused by not freed keymap
object.

As long as we can have only one keymap from buffer at a time,
store each keymap after a new context is created as unique_ptr
that will be freed with a custom deleter on either dtor or when
another keymap is created.

Bug: 1120270
Change-Id: I741a6cb74713424f54dd858f84c0fec413cd6c44
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367938
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Maksim Sisov (GMT+3) <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#800708}