AppCache: Remove 304 resource response handling.
When Chrome receives a 304 for a resource stored in AppCache, we do not
update the cache's headers with the 304 headers.

We had a fix for this issue under development, gated by a feature flag.
While the fix was very close to the finish line, AppCache's deprecation
and removal means we can't justify finishing the fix.

This CL removes the feature flag and the code gated by it. This is
intended to make it easier to maintain our AppCache implementation until
we're ready to delete it from the tree.

This CL should not cause any behavior changes for Chrome instances
without manually modified feature flags.

Bug: 989611
Change-Id: I9f0c98fbbc59dab527db8159eb6f691398480c2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2360556
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: enne <enne@chromium.org>
Commit-Queue: Victor Costan <pwnall@chromium.org>
Auto-Submit: Victor Costan <pwnall@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800735}