Production implementation of encryption/decryption.
Fake-* is removed, as well as virtual interface: we are using openssl and AEAD directly.
Also, decryption is moved out from Chrome, since for now we only use it in test.
Once we add it to the server, it will be placed elsewhere and probably wrapped in Java class.

Bug: b:153649905
Change-Id: Ib6383efa8f6b1e993e351e03f2f747043c6cb521
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363868
Commit-Queue: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: Zach Trudo <zatrudo@google.com>
Cr-Commit-Position: refs/heads/master@{#800750}