Store CoreAccountInfo in ExtensionTokenKey
Skipping the account chooser will require the GetAuthToken function to
retrieve the gaia id of the account currently in use. Storing the
CoreAccountInfo will allow the function to always be able to retrieve
the gaia id for comparison in the scenario that CoreAccountId is an
email.

This CL replaces the CoreAccountId with CoreAccountInfo in the
ExtensionTokenKey. Additionally, relevant locations of the existing code
have been modified to use the CoreAccountId stored in CoreAccountInfo to
account for this replacement.

Bug: 1100535
Change-Id: Ie72f437aecdbfe6b98a1c5cf0caae3501f42baa7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2366033
Reviewed-by: John Lee <johntlee@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Commit-Queue: William Lin <williamlin@google.com>
Cr-Commit-Position: refs/heads/master@{#800754}