[Cryptids] Add logging of render event
This CL logs a successful cryptid rendering when the asset is
successfully fetched and injected, and invokes writing the timestamp to
the relevant pref. It also adds logging for failed asset fetches, and as
a baseline logs instances where the conditions for rendering were not
satisfied.

I considered using a Histogram for this logging; however, I do intend to
look at the action sequences following a successful render, so I thought
a User Action was more appropriate.

Bug: 1120348
Change-Id: Ib9a536aa928591a2105cdaaca4e054b2c2a01ef8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367642
Reviewed-by: Steven Holte <holte@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Tommy Martino <tmartino@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800767}