[webauthn] Initial support for large blobs
This is an initial patch to add support for authenticator large blobs.
When a chrome extension creates a resident credential on an
authenticator that supports largeBlobKey, a key is generated and stored
alongside the credential.

At the moment this does nothing in the real world since there are no
hardware implementations for the authenticator extension.

See https://w3c.github.io/webauthn/#sctn-large-blob-extension

Bug: 1114875
Change-Id: Ib0791e143ab78d9f09461ee0a3226fb9c23f75cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2353150
Commit-Queue: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Martin Kreichgauer <martinkr@google.com>
Cr-Commit-Position: refs/heads/master@{#800775}