Allow GetAuthToken to skip the account chooser
Unbundled consent will allow GetAuthToken to make API calls to the
server to automatically select an account to request permissions from.
This account can be one specified by the extension or one that
previously granted permissions to the extension. This can be useful for
users so they don't need to choose or remember the "correct" account in
the account chooser. This CL adds this feature of skipping the account
chooser and puts it behind a flag to enable it.

Bug: 1100535
Change-Id: I859784b5d3a533144c48802845b3317ce0e01935
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2363879
Commit-Queue: William Lin <williamlin@google.com>
Reviewed-by: Devlin <rdevlin.cronin@chromium.org>
Reviewed-by: Alex Ilin <alexilin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800794}