viz: Changing color usage invalidates the whole frame
When the root contents color usage changes between frames
the renderer might reshape the output surface, increasing
the damage to the whole frame.

If SurfaceAggregator ignores quads outside the damage we
might end up with a renderer that can't draw the full frame
after reshaping the surface.

This CL increases the damage in SurfaceAggragtor to the full
frame when the color usage changes.

Bug: b/158777601
Test: Youtube HDR video, no artifacts when it starts playing.
Change-Id: Ibce88f6f4375567c1bbab1761ab7aea53f84c226
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2369441
Commit-Queue: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Daniele Castagna <dcastagna@chromium.org>
Reviewed-by: Miguel Casas <mcasas@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800807}