BackForwardCache: Emit 0 samples for key metrics for bfcache
Now no samples are emitted for regular VOLT metrics after the page is
restored from the back-forward cache. This means that we will miss a lot
of metrics for history navigations after we launch back-forward cache.
As metrics for history navigations tend to be better figures than other
navigations (e.g., due to network cache), the average of such metrics
values will become worse and might seem regression if we don't take any
actions.

To mitigate this issue, we plan to emit 0 samples for such key metrics
for back-forward navigations. This is implemented behind the flag so
far, and we will enable this by default when we reach the conclusion how
to adjust them.

Bug: 1014174
Bug: 1121066
Change-Id: I04bf9646167aaf4c7241496a12ad66be3fe92681
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2354369
Commit-Queue: Hajime Hoshi <hajimehoshi@chromium.org>
Reviewed-by: Nicolás Peña Moreno <npm@chromium.org>
Reviewed-by: Alexander Timin <altimin@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Cr-Commit-Position: refs/heads/master@{#800957}