Ozone/X11:custom_frame_pref_default value from GetCustomFramePrefDefault
non-Ozone/X11 uses the GetCustomFramePrefDefault() method to determine
whether a custom frame needs to be used. In contrast, Ozone/X11
always set that to false. That looks like an inconsistency between
twos.

Thus, use that method instead.

PS: I had to remove constexpr from the kX11PlatformProperties
to be able to change the value. It doesn't seem to be possible
to call the GetCustomFramePrefDefault method from initializer
as long as that method isn't constexpr and cannot be changed to
that either.

Bug: 789065
Change-Id: Ibd5e7387ea6d3e473f217638f203f74ae42c530f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2372462
Reviewed-by: Thomas Anderson <thomasanderson@chromium.org>
Commit-Queue: Maksim Sisov (GMT+3) <msisov@igalia.com>
Cr-Commit-Position: refs/heads/master@{#801048}