[scroll-animations] Allow null source in ScrollTimeline constructor
The spec was recently updated to differentiate between the source
being missing and the source being null. The former means that the
source should default to document.scrollingElement, and the latter
means that there is no source.

Note: In the CSS API, this translates to 'auto' and 'none',
respectively. Support for this has already landed in a previous CL.

Bug: 1100907
Change-Id: I1436b16d06b7a7ae0de55e2706a7b5d87c0290b5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2367837
Reviewed-by: Kevin Ellis <kevers@chromium.org>
Commit-Queue: Anders Hartvoll Ruud <andruud@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801085}