Update root style environment variables for foldable vars
Changes to the window segments happens in the content layer as part of
the visual properties cascade. As such the current machinery to compute
and update the associated CSS environment variables is not triggered
in scenarios like page reload.

Instead of setting the variables on each frame, update the root
environment variables instance, which is used by child
DocumentStyleEnvironmentVariables instances when resolving.

Bug: 1121068
Change-Id: I0a46b41957c04708f4bd255b87a00c154477d893
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2371684
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Daniel Libby <dlibby@microsoft.com>
Cr-Commit-Position: refs/heads/master@{#801113}