Update Android GPU watchdog init and restart factors for finch.
Set kRestartFactorFinch and kInitFactorFinch to 4 for Android. This will
set the initialization and restart timeout to 60 seconds (4*15 seconds)
for kGpuWatchdogV2NewTimeout finch on Android.

For kGpuChannelTimeoutInSeconds in the browser process on Android, use
the timeout constant from the watchdog instead of a hard-coded value.

Bug: 1120145
Change-Id: I32e39c5b005a9957cf1d4fff58c4aabcb3964397
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2282782
Reviewed-by: Bo <boliu@chromium.org>
Reviewed-by: Zhenyao Mo <zmo@chromium.org>
Commit-Queue: Maggie Chen <magchen@chromium.org>
Cr-Commit-Position: refs/heads/master@{#801126}